fix(variant-study): copy with path

+++ antarest/study/storage/variantstudy/variant_study_service.py
@@ -900,6 +900,7 @@ class VariantStudyService(AbstractStorageService[VariantStudy]):
         dest_name: str,
         groups: Sequence[str],
         with_outputs: bool = False,
+        destination_path: str = "",
     ) -> RawStudy:
         """
         Create a new variant study by copying a reference study.
@@ -909,21 +910,22 @@ class VariantStudyService(AbstractStorageService[VariantStudy]):
             dest_name: The name for the destination study.
             groups: A list of groups to assign to the destination study.
             with_outputs: Indicates whether to copy the outputs as well.
+            destination_path: Path where the destination study will be stored. If not specified, the destination path will be the same as the source study.
 
         Returns:
             The newly created study.
         """
 
-        dest_study = self.raw_study_service.build_raw_study(dest_name, groups, src_study)
+        dest_study = self.raw_study_service.build_raw_study(dest_name, groups, src_study, destination_path)
 
         file_study = self.get_raw(metadata=src_study)
 
         src_path = file_study.config.path
         dest_path = dest_study.path
-        shutil.copytree(src_path, dest_path)
+        shutil.copytree(str(src_path), str(dest_path))
 
-        if with_outputs:
-            src_path = cast(Path, file_study.config.output_path)
+        src_path = cast(Path, file_study.config.output_path)
+        if with_outputs and src_path.exists():
             dest_path = Path(dest_study.path) / OUTPUT_RELATIVE_PATH
             shutil.copytree(src_path, dest_path)
